requirements.txt
discord.py==2.3.2
aiohttp==3.9.5
python-dotenv==1.0.1

disnake==2.9.1
